pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dh-tool"
version = "1.0.0"
authors = [
  { name="dohyun", email="doh0106@naver.com" },
]
description = "A small example package"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    'elasticsearch',
    'matplotlib',
    'numpy',
    'openai',
    'openpyxl',
    'pandas',
    'python_box',
    'scikit_learn',
    'seaborn',
    'tqdm'
]

[project.urls]
Homepage = "https://github.com/pypa/sampleproject"
Issues = "https://github.com/pypa/sampleproject/issues"